contributor metadata in metanorma extensions: https://github.com/metanorma/metanorma-nist/issues/277
