+++ lib/metanorma/standoc/cleanup.rb
@@ -71,6 +71,7 @@ module Metanorma
         mathml_cleanup(xmldoc)
         script_cleanup(xmldoc)
         docidentifier_cleanup(xmldoc) # feeds: bibdata_cleanup
+        ext_contributor_cleanup(xmldoc) # feeds: bibdata_cleanup
         bibdata_cleanup(xmldoc)
         svgmap_cleanup(xmldoc) # feeds: img_cleanup
         boilerplate_cleanup(xmldoc)
